Add support for enum-like values
Added support for "RemapExpressions" and "FKRemapExpressions. These
are useful when a user has an integer column that represents a textual
value. A RemapExpression is when the user provides the map from the
integer value to a textual representation. An FKRemapExpression is
when that lookup is done via an FK related table.